+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [0.71.0]
 
+* Clear output channel before running the "Logtalk: Run Project Testers" and "Logtalk: Run Project Doclets" commands
 * Add "Split in individual directives" refactoring support
 * Add "Extract to Logtalk entity" refactoring support